scrapers/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
pydantic>=2.0.0
flask>=2.0.0
flask-cors>=4.0.0
playwright>=1.40.0
aiohttp>=3.8.0
asyncio
requests>=2.28.0
beautifulsoup4>=4.11.0
selenium>=4.0.0
pyotp>=2.8.0
python-dotenv>=1.0.0
